requirements.txt
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3
google-generativeai==0.8.5
google-api-core==2.25.1
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-api-python-client==2.181.0
protobuf==5.29.5
requests==2.32.5
